% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScan.R
\name{addScan}
\alias{addScan}
\title{Additive ANOVA genome scan}
\usage{
addScan(pheno, snp, mc.cores = 1L)
}
\arguments{
\item{pheno}{An object of class \code{\link{qtcatPheno}}.}

\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{mc.cores}{Number of cores for parallelising. Theoretical maximum is
\code{'B'}. For details see \code{\link[parallel]{mclapply}}.}
}
\description{
Additive ANOVA genome scan, scanning each marker.
}
