% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiScan.R
\name{epiScan}
\alias{epiScan}
\title{Epistatic ANOVA genome scan}
\usage{
epiScan(pheno, snp, iInx, mc.cores = 1L)
}
\arguments{
\item{pheno}{An object of class \code{\link{qtcatPheno}}.}

\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{iInx}{An object of class \code{\link{iqtcatInx}}.}

\item{mc.cores}{Number of cores for parallelising. Theoretical maximum is
\code{'B'}. For details see \code{\link[parallel]{mclapply}}.}
}
\description{
Epistatic ANOVA genome scan, scanning all two marker interactions.
}
