% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{iClarans}
\alias{iClarans}
\title{A Clustering Algorithm based on Randomized Search with linkage 
correlation distance}
\usage{
iClarans(snp, iInx, k, maxNeigbours = 100, nLocal = 10, mc.cores = 1)
}
\arguments{
\item{snp}{A object of class \linkS4class{snpMatrix}.}

\item{iInx}{...}

\item{k}{positive integer specifying the number of clusters, less than 
the number of observations.}

\item{maxNeigbours}{positive integer specifying the maximum number of 
randomized searches.}

\item{nLocal}{positive integer specifying the number of optimisation runs.
Columns have to be similar to \code{snp}.}

\item{mc.cores}{number of cores for parallel computing. See \code{mclapply} 
in package parallel for details.}
}
\description{
Partitioning (clustering) into k clusters "around medoids" by 
randomized search. 1-abs(cor) is used as distance.
}
