% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{iDistCor}
\alias{iDistCor}
\title{Correlation distance interactions}
\usage{
iDistCor(snp, iInx)
}
\arguments{
\item{snp}{A object of class \linkS4class{snpMatrix}.}

\item{iInx}{...}
}
\description{
Correlation distance interactions
}
