% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{iIdenticals}
\alias{iIdenticals}
\title{Internal function}
\usage{
iIdenticals(snp, iInx, mc.cores = 1L)
}
\arguments{
\item{snp}{A object of class \linkS4class{snpMatrix}.}

\item{iInx}{...}

\item{mc.cores}{number of cores for parallel computing. See \code{mclapply}}
}
\description{
cluster at distance zero in hybclustX.cor interactions
}
