% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtcatPackage.R
\docType{package}
\name{iqtcat-package}
\alias{iqtcat-package}
\alias{iqtcat}
\title{Intercation quantitative trait cluster association test}
\description{
Intercation quantitative trait cluster association test
}
\author{
Jonas Klasen
}
