% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{iqtcatClust}
\alias{iqtcatClust}
\title{A three step Clustering Algorithm 
correlation distance}
\usage{
iqtcatClust(snp, iInx, k, identicals = TRUE, maxNeigbours = 100,
  nLocal = 10, method = "complete", mc.cores = 1, trace = FALSE,
  ...)
}
\arguments{
\item{snp}{A object of class \linkS4class{snpMatrix}.}

\item{iInx}{...}

\item{k}{positive integer specifying the number of clusters, less than 
the number of observations.}

\item{identicals}{lgical, if zero clustering ...}

\item{maxNeigbours}{positive integer specifying the maximum number of 
randomized searches.}

\item{nLocal}{positive integer specifying the number of optimisation runs.
Columns have to be similar to \code{snp}.}

\item{method}{see hclust}

\item{mc.cores}{number of cores for parallel computing. See \code{mclapply} 
in package parallel for details.}

\item{trace}{...}

\item{...}{...}
}
\description{
Dendrogram for big data.
}
