% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{iqtcatInx}
\alias{iqtcatInx}
\title{Index of inteacting loci}
\usage{
iqtcatInx(snp, minDist = 10000)
}
\arguments{
\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{minDist}{minimal physical distence}
}
\description{
Index of inteacting loci
}
