% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiScan.R
\name{iSinglTests}
\alias{iSinglTests}
\title{A two marker interaction  ANOVA}
\usage{
iSinglTests(i, y, snp, iInx, x.des, family, test = c("LRT", "F"))
}
\arguments{
\item{i}{...}

\item{y}{...}

\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{iInx}{An object of class \code{\link{iqtcatInx}}.}

\item{x.des}{...}
}
\description{
Internal two marker interaction  ANOVA
}
\keyword{internal}
