% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScan.R
\name{plot.addScan}
\alias{plot.addScan}
\title{Plot of significance of marker-phenotype association.}
\usage{
\method{plot}{addScan}(x, xlab = "Chromosomes",
  ylab = expression(-log[10](italic(p))), col.axis = NULL, ...)
}
\arguments{
\item{x}{A data.frame object of class \code{\link{addScan}}.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{col.axis}{Colors for axis line, tick marks, and title respectively.}

\item{...}{Other graphical parameters may also be passed as arguments to this function.}
}
\description{
Plot of significance of marker-phenotype association at their
genome positions.
}
