% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiScan.R
\name{plot.epiScan}
\alias{plot.epiScan}
\title{Plot of significance of (marker by marker)-phenotype association.}
\usage{
\method{plot}{epiScan}(x, threshold = 0.05, col.axis = "black",
  col.lab = "black", ...)
}
\arguments{
\item{x}{a data.frame object of class \code{\link{epiScan}}.}

\item{threshold}{a significance threshold, used for coloring of plot.}

\item{col.axis}{the color to be used for axis annotation. Defaults to "black".}

\item{col.lab}{the color to be used for x and y labels. Defaults to "black".}

\item{...}{other graphical parameters may also be passed as arguments to this function.}
}
\description{
Plot of significance of (marker by marker)-phenotype association at their
genome positions.
}
