% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addScan.R
\name{singleTests}
\alias{singleTests}
\title{Single marker ANOVA}
\usage{
singleTests(i, y, snp, x.des, family, test = c("LRT", "F"))
}
\arguments{
\item{i}{...}

\item{y}{...}

\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{x.des}{...}

\item{family}{...}

\item{test}{...}
}
\description{
Internal singele marker ANOVA
}
\keyword{internal}
