% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{test.iInx}
\alias{test.iInx}
\title{Test index of inteacting markers}
\usage{
test.iInx(snp, iInx, minDist = 10000)
}
\arguments{
\item{snp}{An object of S4 class \linkS4class{snpMatrix}.}

\item{iInx}{An object of class \code{\link{iqtcatInx}}.}

\item{minDist}{minimal physical distence}
}
\description{
Test if inteacting markers have all four homozygous combination, 
if not, they are removed.
}
